schema definition, trait
