+++ runtime/serde/common/src/aws/smithy/kotlin/runtime/serde/schema/trait/Trait.kt
@@ -1,0 +1,11 @@
+/*
+ * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ */
+package aws.smithy.kotlin.runtime.serde.schema.trait
+
+import aws.smithy.kotlin.runtime.serde.schema.ShapeId
+
+public interface Trait {
+    public val id: ShapeId
+}
